requirements.txt
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.5.0
clip==1.0
ftfy==5.9
numpy==1.20.1
Pillow==8.1.1
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
six==1.15.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.58.0
typing-extensions==3.7.4.3
wcwidth==0.2.5
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
bert-for-tf2==0.14.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
flatbuffers==1.12
gast==0.2.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==4.4.0
joblib==1.0.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.4
params-flow==0.8.2
Pillow==8.2.0
protobuf==3.17.1
py-params==0.10.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.1
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
